#27 factor out some code into style.py and Vpython.py
